requirements.txt
alembic==1.11.1
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.3.2
flask-restx==1.1.0
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pydash==7.0.3
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
sentry-sdk==1.21.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.1
urllib3==2.0.2
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
